Introduce common-plugin module
