+++ gradle-plugin/build.gradle.kts
@@ -1,22 +1,22 @@
 plugins {
-	`kotlin-dsl`
-	`java-gradle-plugin`
-	`maven-publish`
-	alias(libs.plugins.dokka)
+    `kotlin-dsl`
+    `java-gradle-plugin`
+    `maven-publish`
+    alias(libs.plugins.dokka)
 }
 
 repositories {
-	mavenCentral()
+    mavenCentral()
 }
 
 val javaVersion = providers.gradleProperty("javaVersion").map(String::toInt).get()
 
 kotlin {
-	jvmToolchain(javaVersion)
+    jvmToolchain(javaVersion)
 }
 
 tasks.test {
-	useJUnitPlatform()
+    useJUnitPlatform()
 }
 
 repositories {
@@ -26,26 +26,21 @@ repositories {
 
 dependencies {
     implementation(project(":core"))
-    implementation(libs.kotlinCompilerEmbeddable)
-    implementation(libs.kotlinStdlib)
-    implementation(libs.asm)
-    implementation(libs.asmTree)
-    testImplementation(gradleTestKit())
-    testImplementation(libs.bundles.testCommon)
+    implementation(project(":common-plugin"))
 }
 
 gradlePlugin {
-	plugins {
-		create("tapikGradlePlugin") {
-			id = "dev.akif.tapik.gradle"
-			implementationClass = "dev.akif.tapik.gradle.TapikGradlePlugin"
-			displayName = "Tapik Gradle Plugin"
-			description = "Tapik code generator Gradle plugin"
-			tags.set(listOf("tapik", "codegen"))
-		}
-	}
+    plugins {
+        create("tapikGradlePlugin") {
+            id = "dev.akif.tapik.plugin.gradle"
+            implementationClass = "dev.akif.tapik.plugin.gradle.TapikGradlePlugin"
+            displayName = "Tapik Gradle Plugin"
+            description = "Tapik code generator Gradle plugin"
+            tags.set(listOf("tapik", "codegen"))
+        }
+    }
 }
 
 publishing {
-	publications { /* created automatically by java-gradle-plugin; keep for mavenLocal */ }
+    publications { /* created automatically by java-gradle-plugin; keep for mavenLocal */ }
 }
