--- gradle-plugin/settings.gradle.kts
@@ -1,17 +1,0 @@
-pluginManagement {
-	repositories {
-		mavenLocal()
-		gradlePluginPortal()
-		mavenCentral()
-	}
-}
-
-dependencyResolutionManagement {
-	@Suppress("UnstableApiUsage")
-	repositories {
-		mavenCentral()
-	}
-}
-
-rootProject.name = "gradle-plugin"
-
